! Test feature
This is an adaptive solution to the original issue. If a client sends data in chunks, there is a chance the end is not seen if we exit too early. With that being said, there is a timer that starts upon last receiving data. If no new data has been received over a given amount of time, we determine that there is no more to send, breaking out of our loop.